Split out direct, timer and converter related tests out of core tests #2362
