+++ integration-test-groups/foundation/pom.xml
@@ -34,8 +34,12 @@
     <modules>
         <!-- extensions a..z; do not remove this comment, it is important when sorting via  mvn process-resources -Pformat -->
         <module>core</module>
+        <module>core-annotations</module>
         <module>customized-log-component</module>
+        <module>direct</module>
         <module>log</module>
+        <module>timer</module>
+        <module>type-converter</module>
     </modules>
 
 </project>
