+++ integration-test-groups/foundation/timer/src/main/java/org/apache/camel/quarkus/component/timer/it/TimerProducers.java
@@ -1,0 +1,35 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.camel.quarkus.component.timer.it;
+
+import javax.enterprise.context.ApplicationScoped;
+
+import org.apache.camel.builder.LambdaRouteBuilder;
+import org.jboss.logging.Logger;
+
+@ApplicationScoped
+public class TimerProducers {
+
+    private static final Logger LOG = Logger.getLogger(TimerProducers.class);
+
+    @javax.enterprise.inject.Produces
+    public LambdaRouteBuilder lambdaRoute() {
+        return rb -> rb.from("timer:bar").routeId("bar")
+                .process(e -> LOG.info("Hello from timer:bar"));
+    }
+
+}
